HazReg.jl

All docstrings from the package

HazReg.ExponentiatedWeibull — Type

ExponentiatedWeibull(sigma,nu,gamma)

The ExponentiatedWeibull distribution with scale sigma, shape nu and second shape gamma has probability density function

\[f(x; parameters) = ...\]

More details and examples of usage could be provided in this docstring.

Maybe this distribution could simply be constructed from a transformation of the original Weibull ?

References:

Link to my reference so that people understand what it is

HazReg.GeneralizedGamma — Type

GeneralizedGamma(sigma,nu,gamma)

The GeneralizedGamma distribution with scale sigma, shape nu and second shape gamma has probability density function

\[f(x; parameters) = ...\]

More details and examples of usage could be provided in this docstring.

Maybe this distribution could simply be constructed from a transformation of the original Weibull ?

References:

Link to my reference so that people understand what it is

HazReg.LogLogistic — Type

LogLogistic(mu,sigma)

To be described...

\[f(x; parameters) = ...\]

HazReg.PowerGeneralizedWeibull — Type

PowerGeneralizedWeibull(sigma,nu,gamma)

The PowerGeneralizedWeibull distribution with scale sigma, shape nu and second shape gamma has probability density function

\[f(x; parameters) = ...\]

More details and examples of usage could be provided in this docstring.

Maybe this distribution could simply be constructed from a transformation of the original Weibull ?

References:

Link to my reference so that people understand what it is

HazReg.GHMLE — Method

GHMLE(...)

Maximum likelihood estimation in General Hazards models using several parametric baseline hazards

Docs to rewrite....

init : initial point for optimisation step under the parameterisation (log(scale), log(shape1), log(shape2), alpha, beta) for scale-shape1-shape2 models or (mu, log(scale), alpha, beta) for log-location scale models.

times : times to event

status: vital status indicators (true or 1 = observed, false or 0 = censored)

hstr: hazard structure. No covariates ("baseline"), AFT model ("AFT"), PH model ("PH"), AH model ("AH"), GH model ("GH").

dist: baseline hazard distribution, on the form of a type <: Distributions.ContinuousUnivariateDistribution. Tested with: - LogNormal - LogLogistic - Weibull - Gamma - ExponentiatedWeibull - GeneralizedGamma - PowerGeneralizedWeibull.

des: design matrix for hazard-level effects

des_t: design matrix for time-level effects (it is recommended not to use splines here)

method: one of NelderMead(), Newton(), LBFGS(), ConjugateGradient() or GradientDescent() or any other method taken by Optim.optimize().

maxit: maximum number of iterations of the optimization routine.

References:

Link to my reference so that people understand what it is

HazReg.simGH — Method

simGH(...)

Details...

Docs to rewrite....

simGH function: Function to simulate times to event from a model with AH, AFT, PH, GH structures for different parametric baseline hazards. Distributions: LogNormal, LogLogistic, GenGamma, GGamma, Weibull, PGW, EW. See: https://github.com/FJRubio67/HazReg ––––––––––––––––––––––––––––––––––––––––––––––––– =#

#= seed : seed for simulation n : sample size (number of individuals) beta : regression parameters multiplying the hazard for the GH model or the regression parameters for AFT and PH models alpha : regression parameters multiplying the time scale for the GH model or the regression parameters for the AH model des : Design matrix for the GH model (hazard scale) or design matrix for AFT and PH models des_t : Design matrix for the GH model (time scale) or design matrix for the AH model hstr : hazard structure (AH, AFT, PH, GH) dist: distribution passed as an already instanciated distribution from Distributions.jl baseline : baseline hazard distribution

Returns a vector containing the simulated times to event

References:

Link to my reference so that people understand what it is